fix: SearchWordResponseDto -> libraryId 변수 타입 Long으로 변경

+++ app/src/main/java/com/example/avocado_android/domain/model/response/main/SearchWordResponseDto.kt
@@ -6,7 +6,7 @@ import com.example.avocado_android.utils.enunm.SearchRequestType
 data class SearchWordResponseDto(
     val isSuccess: Boolean? = null,
     val isLibraryRegistered: Boolean? = null,
-    val libraryId: Int? = null,
+    val libraryId: Long? = null,
     val characterImgUrl: String? = null,
     val contents: Map<SearchRequestType, String>? = null
 )
